refactor internals
