+++ avaje-jex-freemarker/src/main/java/io/avaje/jex/render/freemarker/FreeMarkerRender.java
@@ -10,7 +10,7 @@ import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import freemarker.template.Version;
 import io.avaje.jex.Context;
-import io.avaje.jex.TemplateRender;
+import io.avaje.jex.spi.TemplateRender;
 import io.avaje.spi.ServiceProvider;
 
 @ServiceProvider
