+++ avaje-jex/src/main/java/io/avaje/jex/core/TemplateManager.java
@@ -1,7 +1,7 @@
 package io.avaje.jex.core;
 
 import io.avaje.jex.Context;
-import io.avaje.jex.TemplateRender;
+import io.avaje.jex.spi.TemplateRender;
 
 import java.util.HashMap;
 import java.util.HashSet;
