+++ avaje-jex/src/main/java/io/avaje/jex/http/HttpResponseException.java
@@ -3,6 +3,10 @@ package io.avaje.jex.http;
 import java.util.Collections;
 import java.util.Map;
 
+/**
+ * Throwing an uncaught {@code HttpResponseException} will interrupt http processing and set the
+ * status code and response body with the given message
+ */
 public class HttpResponseException extends RuntimeException {
 
   private final int status;
