+++ avaje-jex/src/main/java/io/avaje/jex/spi/JexExtension.java
@@ -1,6 +1,5 @@
 package io.avaje.jex.spi;
 
-import io.avaje.jex.TemplateRender;
 import io.avaje.spi.Service;
 
 @Service
